+++ pkg/query.go
@@ -57,8 +57,8 @@ type queryModel struct {
 	FillMode    string   `json:"fillMode"`
 }
 
-func (qc *queryConfigStruct) fetchData(ctx context.Context, config *pluginConfig, password string, privateKey string) (result DataQueryResult, err error) {
-	connectionString := getConnectionString(config, password, privateKey)
+func (qc *queryConfigStruct) fetchData(ctx context.Context, config *pluginConfig, password string, privateKey string, token string) (result DataQueryResult, err error) {
+	connectionString := getConnectionString(config, password, privateKey, token)
 
 	db, err := sql.Open("snowflake", connectionString)
 	if err != nil {
@@ -180,7 +180,7 @@ func (qc *queryConfigStruct) transformQueryResult(columnTypes []*sql.ColumnType,
 	return values, nil
 }
 
-func (td *SnowflakeDatasource) query(ctx context.Context, dataQuery backend.DataQuery, config pluginConfig, password string, privateKey string) (response backend.DataResponse) {
+func (td *SnowflakeDatasource) query(ctx context.Context, dataQuery backend.DataQuery, config pluginConfig, password string, privateKey string, token string) (response backend.DataResponse) {
 	var qm queryModel
 	err := json.Unmarshal(dataQuery.JSON, &qm)
 	if err != nil {
@@ -219,7 +219,7 @@ func (td *SnowflakeDatasource) query(ctx context.Context, dataQuery backend.Data
 	queryConfig.FinalQuery = strings.TrimSuffix(strings.TrimSpace(queryConfig.FinalQuery), ";")
 
 	frame := data.NewFrame("")
-	dataResponse, err := queryConfig.fetchData(ctx, &config, password, privateKey)
+	dataResponse, err := queryConfig.fetchData(ctx, &config, password, privateKey, token)
 	if err != nil {
 		response.Error = err
 		return response