+++ src/types.ts
@@ -28,9 +28,12 @@ export interface SnowflakeOptions extends DataSourceJsonData {
   database?: string;
   schema?: string;
   extraConfig?: string;
+  /** @deprecated use authMethod */
   basicAuth: boolean;
   maxChunkDownloadWorkers?: string;
   customJSONDecoderEnabled: boolean;
+  authMethod: string;
+  clientId?: string;
 }
 
 /**
@@ -39,4 +42,6 @@ export interface SnowflakeOptions extends DataSourceJsonData {
 export interface SnowflakeSecureOptions {
   password?: string;
   privateKey?: string;
+  clientSecret?: string;
+  tokenEndpoint?: string;
 }